Products: Secured backend from unathorized to manage products

+++ products/views.py
@@ -3,6 +3,7 @@ from django.shortcuts import (
     HttpResponse, reverse
 )
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Category, Product
 from .forms import ProductForm
 
@@ -104,8 +105,13 @@ def remove_from_package(request, item_id):
         return HttpResponse(status=500)
 
 
+@login_required
 def add_product(request):
     """ Add a product. """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect('home')
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -128,8 +134,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect('home')
+
     product = get_object_or_404(Product, id=product_id)
 
     if request.method == 'POST':
@@ -156,8 +167,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a product """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect('home')
+
     product = get_object_or_404(Product, id=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
